sam-app/lambda/requirements.txt
line-bot-sdk
langchain
langchain-openai
langchain-pinecone
